% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_approx.R
\name{add_approx}
\alias{add_approx}
\title{Add deterministic approximation line}
\usage{
add_approx(path)
}
\arguments{
\item{path}{Path to the approximation data file (this should be an RDS file)}
}
\description{
Adds lines for a deterministic simulation approximating the stochastic
simulation, on top of the trait density plot through time.
}
\note{
See the \code{speciome-approx} repository and \code{speciomx} package
for how to simulate the deterministic model.

https://github.com/rscherrer/speciome-approx
}
\examples{

\dontrun{

root <- system.file("extdata", "sim-example", package = "speciomer")
approx <- file.path(root, "approx.rds")
trait_data <- read_individuals(root, "traits", ncols = 3)
trait_plot <- plot_traits(trait_data)
trait_plot + add_approx(approx)

}

}
