% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_line.R
\name{add_line}
\alias{add_line}
\title{Add a line to a ggplot}
\usage{
add_line(data, y)
}
\arguments{
\item{data}{A data frame}

\item{y}{The name of the variable to plot as a line (as a string)}
}
\description{
Add a line to a ggplot
}
\details{
This function maps variable \code{y} to the \code{y} aesthetic of a
\code{geom_line} and adds it to a preexisting ggplot.
}
\note{
\code{y} cannot be "y".
}
\examples{

data <- tibble::tibble(x = 1:3, z = 3:1)

ggplot2::ggplot(data, ggplot2::aes(x = x, y = x)) +
  ggplot2::geom_line() +
  add_line(data, "z")

}
