% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble_speciome.R
\name{as_tibble_speciome}
\alias{as_tibble_speciome}
\title{Turn a data vector into a tibble}
\usage{
as_tibble_speciome(x, ncol = 1)
}
\arguments{
\item{x}{Vector of data values}

\item{ncol}{Number of columns by which to split the vector. Use a negative
value to duplicate the vector into a longer one-column tibble instead of
splitting it.}
}
\value{
A tibble containing the simulation data
}
\description{
Reshapes a speciome data vector into a matrix and converts it into a tibble.
}
\seealso{
`read_speciome`
}
