% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecotype_colors.R
\name{ecotype_colors}
\alias{ecotype_colors}
\title{Ecotype colors}
\usage{
ecotype_colors()
}
\value{
A vector of strings
}
\description{
Returns the colors to use when plotting the two ecotypes. Basically white
and black.
}
\examples{

ecotype_colors()

}
