% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locus_data.R
\name{get_locus_data}
\alias{get_locus_data}
\title{Extract data for one locus}
\usage{
get_locus_data(i, data, variables, add_time = FALSE)
}
\arguments{
\item{i}{Index of the locus to extract data for}

\item{data}{Locus-wise data frame trough time (see \code{?read_loci})}

\item{variables}{String vector with the names of the columns to extract}

\item{add_time}{Whether or not to extract the "time" column as well}
}
\value{
A tibble
}
\description{
Extracts a tibble with the data for one locus from a locus-wise dataset
through time.
}
\examples{

root <- system.file("extdata", "sim-example", package = "speciomer")
data <- read_loci(root, "Fst")
get_locus_data(1, data, "Fst")

}
\seealso{
\code{read_loci}
}
