% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locus_data_n.R
\name{get_locus_data_n}
\alias{get_locus_data_n}
\title{Extract data for one locus (nested)}
\usage{
get_locus_data_n(i, data_n, variables, add_time = FALSE)
}
\arguments{
\item{i}{Index of the locus to extract data for}

\item{data_n}{Locus-wise data frame trough time (see \code{?read_loci}),
nested by locus (use \code{dplyr::group_by} and \code{tidyr::nest} to get
that)}

\item{variables}{String vector with the names of the columns to extract}

\item{add_time}{Whether or not to extract the "time" column as well}
}
\value{
A tibble
}
\description{
Same as \code{get_locus_data} but extracting from a data frame nested by
locus (which is much faster).
}
\examples{

root <- system.file("extdata", "sim-example", package = "speciomer")
data <- read_loci(root, "Fst")
data <- data \%>\% dplyr::group_by(locus) \%>\% tidyr::nest()
get_locus_data_n(1, data, "Fst")

}
\seealso{
\code{read_loci}, \code{dplyr::group_by}, \code{tidyr::nest}
}
