% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave_all.R
\name{ggsave_all}
\alias{ggsave_all}
\title{Save many ggplots}
\usage{
ggsave_all(plots, names, width, height, dpi = 300)
}
\arguments{
\item{plots}{A list of ggplot objects}

\item{names}{A vector of file names (where to save each plot)}

\item{width, height, dpi}{Single values to be passed to \code{ggsave}}
}
\description{
Loops through ggplots and their respective figure names and saves them.
}
\note{
The same \code{width}, \code{height} and \code{dpi} is used for all
the \code{plots}
}
\examples{

\dontrun{

plot1 <- ggplot2::quickplot()
plot2 <- plot1
ggsave_all(list(plot1, plot2), c("p1.png", "p2.png"), width = 3, height = 3)

}

}
\seealso{
\code{ggplot2::ggsave}
}
