% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_variable_names.R
\name{interpret_variable_names}
\alias{interpret_variable_names}
\title{Interpret variable names}
\usage{
interpret_variable_names(x, type)
}
\arguments{
\item{x}{Vector of names of variables}

\item{type}{The label to add (either of "locus", "edge", "individual" or "trait")}
}
\value{
A vector of strings
}
\description{
Adds labels to variable names that are provided without labels
}
\details{
The function adds the specified label to the variable names. For example,
"Fst" becomes "locus_Fst" if \code{type = "locus"}. It will not add a label
if it detects that the label is already there, i.e. "locus_Fst" will stay "locus_Fst".
}
\note{
In the case of \code{type = "trait"} some variable names will be considered
exceptions to which not to add a label. These are the variables that are saved
per time point and/or per ecotype or habitat but not per trait
("EI", "RI", "SI", "population_sizes", "ecotype_population_sizes" and "habitat_resources")
}
\examples{

interpret_variable_names(c("Fst", "Qst"), "locus")
interpret_variable_names(c("corbreed", "corfreq"), "edge")

}
