% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_degrees.R
\name{plot_degrees}
\alias{plot_degrees}
\title{Plot locus degrees}
\usage{
plot_degrees(arch, xaxis = "location", rm_x = TRUE)
}
\arguments{
\item{arch}{A data frame containing locus-wise genetic architecture
data (see \code{?read_architecture})}

\item{xaxis}{What to show on the x-axis. Either of "location" for genomic
location (continuous between 0 and 1) or "locus" for locus index.}

\item{rm_x}{Whether to remove the x-axis entirely}
}
\value{
A ggplot
}
\description{
Barplot of the degrees of loci across the genome.
}
\examples{

root <- system.file("extdata", "sim-example", package = "speciomer")
arch <- read_architecture(root)$nodes
plot_degrees(arch)

}
\seealso{
\code{read_architecture}
}
