% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eco_alpha_zoomed.R
\name{plot_eco_alpha_zoomed}
\alias{plot_eco_alpha_zoomed}
\title{Plot ecological average mutational effects at different times}
\usage{
plot_eco_alpha_zoomed(root, root_init)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{root_init}{Path to the initialization simulation folder, i.e. a
simulation that is the equivalent of \code{root}, but simulated only for
a few thousand generations and with a high recording resolution. This
allows us to see what is happening in detail in the early steps.}
}
\value{
A ggplot
}
\description{
Plots three portions of the plot of average mutational effects per locus
and through time, for ecological loci only: over the full
simulation time span, at the very start and around branching time.
}
\details{
The logic is the same as for \code{plot_eco_trait_zoomed}
}
\seealso{
\code{read_individuals}, \code{plot_eco_trait_zoomed}
}
