% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eco_trait_histogram.R
\name{plot_eco_trait_histogram}
\alias{plot_eco_trait_histogram}
\title{Ecological trait histogram}
\usage{
plot_eco_trait_histogram(data, time)
}
\arguments{
\item{data}{A data frame containing individual-wise data with a column named
"trait1" (see \code{read_individuals}).}

\item{time}{The time point at which to plot the distribution}
}
\value{
A ggplot
}
\description{
Plots a histogram of the ecological trait values across the population
at a certain time point.
}
\examples{

root <- system.file("extdata", "sim-example", package = "speciomer")
data <- read_individuals(root, "traits", ncol = 3)
plot_eco_trait_histogram(data, -1000)

}
\seealso{
\code{read_individuals}
}
