% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eco_trait_zoomed.R
\name{plot_eco_trait_zoomed}
\alias{plot_eco_trait_zoomed}
\title{Plot ecological trait density at different times}
\usage{
plot_eco_trait_zoomed(root, root_init, add_rect = FALSE)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{root_init}{Path to the initialization simulation folder, i.e. a
simulation that is the equivalent of \code{root}, but simulated only for
a few thousand generations and with a high recording resolution. This
allows us to see what is happening in detail in the early steps.}

\item{add_rect}{Whether to add a shading rectangle on the left facet to
indicate what part of the plot is zoomed in, in the right facet.}
}
\value{
A ggplot
}
\description{
Plots three portions of the ecological trait density plot: over the full
simulation time span, at the very start and around branching time.
}
\details{
Because the burn-in in \code{root_init} is not as long as in
\code{root}, the starting time is modify so the two simulations match
each other (the two simulations should be identical because they are
simulated with the same seed).
}
\seealso{
\code{read_individuals}
}
