% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eco_zoomed.R
\name{plot_eco_zoomed}
\alias{plot_eco_zoomed}
\title{Plot zoomed ecological dynamics}
\usage{
plot_eco_zoomed(root, root_init)
}
\arguments{
\item{root, root_init}{Arguments to pass to \code{plot_eco_trait_zoomed} and
\code{plot_eco_alpha_zoomed}}
}
\value{
A patchwork
}
\description{
Plots different zoomed windows of the dynamics of the ecological trait
densities, and the corresponding plots of average mutational effects of
ecological loci through time.
}
\seealso{
\code{plot_eco_trait_zoomed}, \code{plot_eco_alpha_zoomed}
}
