% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_full_scan.R
\name{plot_full_scan}
\alias{plot_full_scan}
\title{Plot extended genome scan}
\usage{
plot_full_scan(data, grn, time, trait, trait_data = NULL)
}
\arguments{
\item{data}{A data frame containing locus-specific data (see
\code{?read_loci})}

\item{grn}{A list of two data frames containing the genetic architecture
(the output of \code{read_architecture})}

\item{time}{Which time point to plot the variables for in the genome scans.}

\item{trait}{Which trait to show the network for (1, 2 or 3)}

\item{trait_data}{Data frame with individual-wise trait data
(see \code{?read_indviduals}). If provided, a histogram of ecological trait
values at the specified time point will be added to the patchwork.}
}
\value{
A patchwork
}
\description{
Plots an assemblage of various genome scans at a given time point, locus-
specific elements of the genetic architecture and the gene network for one
of the three traits. Genome scans are shown for allele frequencies, Fst,
Qst, Cst and average mutational effects.
}
\note{
\code{grn} is internally converted to a \code{tbl_graph} before being
passed to \code{plot_network}.
}
\examples{

\dontrun{

root <- system.file("extdata", "sim-example", package = "speciomer")
data <- read_loci(root, c("freq", "Fst", "Qst", "Cst", "alpha"))
grn <- read_architecture(root)
plot_full_scan(data, grn, time = -1000, trait = 1)

}

}
\seealso{
\code{read_architecture}, \code{read_loci}, \code{plot_genome_scan},
\code{plot_barcode}, \code{plot_effect_sizes}, \code{plot_degrees},
\code{plot_network}, \code{read_individuals}
}
