% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_full_scan_from_root.R
\name{plot_full_scan_from_root}
\alias{plot_full_scan_from_root}
\title{Plot extended genome scan (from root folder)}
\usage{
plot_full_scan_from_root(root, times, show_traits = TRUE, show_time = FALSE)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{times}{The time points to plot}

\item{show_time}{Whether to show the time point on each plot}

\item{show_hist}{Whether to show a histogram of ecological trait values too}
}
\value{
A list of ggplots (or a single ggplot if only one time point is
specified)
}
\description{
Plots an extended genome scan with multiple layers at a series of time
points, taking only the root folder of the simulation as input.
}
\details{
Runs \code{plot_full_scan} in the background
}
\examples{

\dontrun{

root <- system.file("extdata", "sim-example", package = "speciomer")
plot_full_scan(root, times = -1000)

}

}
\seealso{
\code{plot_full_scan}, \code{read_loci}, \code{read_architecture}
}
