% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_regression.R
\name{plot_gene_regression}
\alias{plot_gene_regression}
\title{Plot genetic values against genotypes at a certain locus}
\usage{
plot_gene_regression(data, locus)
}
\arguments{
\item{data}{Individual whole genome data with information about individual
ecotypes (\code{read_individual_genomes} with \code{individual_trait =
"ecotypes"}).}

\item{locus}{Index of the locus to plot}
}
\value{
A ggplot
}
\description{
Plots the regression of genetic values against allele counts (genotypes) for
a particular locus, facetted by time point. Genetic values are represented
both by boxplots and dots (individuals are binned by genetic value such that
}
\examples{

root <- system.file("extdata", "sim-indiv-genomes", package = "speciomer")
data <- read_individual_genomes(root, individual_variables = "ecotypes")
plot_gene_regression(data, locus = 1)

}
\seealso{
\code{read_individual_genomes}
}
