% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome_scan.R
\name{plot_genome_scan}
\alias{plot_genome_scan}
\title{Plot a genome scan}
\usage{
plot_genome_scan(
  data,
  variable,
  time,
  xaxis = "location",
  ylab_parsed = NULL,
  rm_x = TRUE
)
}
\arguments{
\item{data}{A data frame containing locus-specific data (see
\code{?read_loci})}

\item{variable}{Name of the variable on the y-axis}

\item{time}{What time point to show the data for}

\item{xaxis}{What to show on the x-axis. Either of "location" for genomic
location (continuous between 0 and 1) or "locus" for locus index.}

\item{ylab_parsed}{Label of the y-axis, parsed}

\item{rm_x}{Whether to remove the x-axis entirely}
}
\value{
A ggplot
}
\description{
Plots a variable across the genome at a certain time point.
}
\examples{

x <- 1 + 1

}
\seealso{
\code{read_loci}
}
