% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loci_tt.R
\name{plot_loci_tt}
\alias{plot_loci_tt}
\title{Plot locus-specific variables through time}
\usage{
plot_loci_tt(data, variable, burnin_bar = TRUE)
}
\arguments{
\item{data}{A data frame with locus-specific data through time (see
\code{?read_loci}).}

\item{variable}{Name of the variable to plot (a string)}

\item{burnin_bar}{Whether to add a vertical bar at time point zero}
}
\value{
A ggplot
}
\description{
Generates a line plot of one variable through time, with one line for each
locus, facetted by trait.
}
\examples{

\dontrun{

root <- system.file("extdata", "sim-example", package = "speciomer")
data <- read_loci(root, "locus_Fst")
data <- data \%>\% dplyr::mutate(trait = recode_traits(trait))
plot_loci_tt(data, "Fst")

}

}
\seealso{
\code{read_loci}
}
