% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Plot a gene network}
\usage{
plot_network(arch, trait, xaxis = "location")
}
\arguments{
\item{arch}{A \code{tbl_graph} containing the genetic architecture (the
output of \code{read_architecture} passed through
\code{tidyrgraph::as_tbl_graph})}

\item{trait}{Name of the trait for which to show the network (as it appears
in \code{arch})}

\item{xaxis}{What to show on the x-axis. Either of "location" for genomic
location (continuous between 0 and 1) or "locus" for locus index.}
}
\value{
A ggraph
}
\description{
Plots a gene regulatory network with a semi-circular layout. Good for
plotting on top of genome scans. Interaction weights are mapped to
edge transparency.
}
\note{
Somehow the \code{ggraph} library needs to be loaded for this function
to produce the expected output (use \code{library(ggraph)}).
}
\examples{

root <- system.file("extdata", "sim-example", package = "speciomer")
arch <- read_architecture(root)
arch <- tidygraph::as_tbl_graph(arch)
plot_network(arch, trait = 1)

}
\seealso{
\code{read_architecture}, \code{tidygraph::as_tbl_graph}
}
