% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outcome.R
\name{plot_outcome}
\alias{plot_outcome}
\title{Plot simulation outcome}
\usage{
plot_outcome(root, which = "full", approx = NULL, ...)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{which}{Which type of outcome to plot? Either of: "full", shows the
trait density plots and all the locus-specific variables through the whole
simulation; "simple", shows only trait density plots and average mutational
effect through time; "burnin", same as "simple" but only during the burn-in
period; "init", same as "simple" but during the first 1,000 generations;
"custom", for custom arguments to be passed to \code{plot_outcome_core}
through \code{...}}

\item{approx}{Path to pass to \code{add_approx}. Leave \code{NULL} for no
overlay approximation plot.}

\item{...}{Further arguments to be passed to \code{plot_outcome_core}}
}
\value{
A patchwork
}
\description{
Plots an overview of a simulation outcome in different ways depending on the
option.
}
\details{
See \code{?plot_outcome_core}
}
\examples{

\dontrun{

root <- system.file("extdata", "sim-example", package = "speciomer")
plot_outcome(root)

}

}
\seealso{
\code{plot_outcome_core}
}
