% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outcome_burnin.R
\name{plot_outcome_burnin}
\alias{plot_outcome_burnin}
\title{Plot the burn-in period}
\usage{
plot_outcome_burnin(root, approx = NULL)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{approx}{Path to pass to \code{add_approx}. Leave \code{NULL} for no
overlay approximation plot.}
}
\value{
A patchwork
}
\description{
Equivalent of \code{plot_outcome_simple} but only shows the burn-in period.
}
\details{
Wrapper around \code{plot_outcome_core}
}
\examples{

\dontrun{

root <- system.file("extdata", "sim-example", package = "speciomer")
plot_outcome_burnin(root)

}

}
\seealso{
\code{plot_outcome_core}
}
