% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outcome_core.R
\name{plot_outcome_core}
\alias{plot_outcome_core}
\title{Plot simulation outcome}
\usage{
plot_outcome_core(
  root,
  from = NULL,
  to = NULL,
  by = 1,
  true_start = NULL,
  full = FALSE,
  approx = NULL
)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{from, to}{Range of time points to plot (defaults to the whole range)}

\item{by}{Frequency of time points to sub-sample when plotting locus-specific
lines. Defaults to 1, all time points are plotted. Use e.g. 100 to plot
every 100 generations.}

\item{true_start}{Shifts the time points by some amount. For example, if an
extra simulation was run with 1,000 generations to record the beginning of
the burn-in with a high resolution of a simulation that would otherwise be
20,000 generation-long in burn-in time, use \code{true_start = -20000}.}

\item{full}{Whether to plot the full figure, i.e. with all locus-specific
plots and not just the trait density plots and the average mutational effect
plots.}

\item{approx}{Path to pass to \code{add_approx}. Leave \code{NULL} for no
overlay approximation plot.}
}
\value{
A patchwork
}
\description{
Plots an overview of a simulation outcome. Shows at least trait density
plots through time, and different locus-level plots depending on the option.
The locus-level plots show allele frequencies, Fst, Qst, Cst and average
mutational effect. Plots are facetted by trait. The time window to plot can
be set.
}
\details{
Details about how the function works.
}
\note{
To keep in mind
}
\examples{

\dontrun{

root <- system.file("extdata", "sim-example", package = "speciomer")
plot_outcome_core(root)

}

}
\seealso{
\code{read_individuals}, \code{read_loci}, \code{read_traits},
\code{plot_traits}, \code{plot_loci_tt}, \code{plot_density},
\code{add_line}, \code{add_approx}
}
