% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outcome_full.R
\name{plot_outcome_full}
\alias{plot_outcome_full}
\title{Plot a full simulation outcome}
\usage{
plot_outcome_full(root, approx = NULL)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{approx}{Path to pass to \code{add_approx}. Leave \code{NULL} for no
overlay approximation plot.}
}
\value{
A patchwork
}
\description{
Plots trait density plots through time as well as locus-specific allele
frequencies, Fst, Qst, Cst and average mutational effects through time,
and histograms showing the distributions of these variables at the last
generation. Variables are shown throughout the whole simulation.
}
\details{
Wrapper around \code{plot_outcome_core}
}
\note{
Locus-specific average mutational effects are subsampled every 1,000
generations to avoid crowding the plots.
}
\examples{

\dontrun{

root <- system.file("extdata", "sim-example", package = "speciomer")
plot_outcome_full(root)

}

}
\seealso{
\code{plot_outcome_core}
}
