% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outcome_init.R
\name{plot_outcome_init}
\alias{plot_outcome_init}
\title{Plot the first generations}
\usage{
plot_outcome_init(root, approx = NULL)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{approx}{Path to pass to \code{add_approx}. Leave \code{NULL} for no
overlay approximation plot.}
}
\value{
A patchwork
}
\description{
Plots the first 1,000 generations of a simulation. Supposed to be used
on a simulation with only 1,000 generations recorded with high resolution.
The start of the x-axis is artificially moved to -20,000 to mimic the early
steps of a longer simulation (without this function we would have had to
simulate 20,000 generations of burn-in with high resolution, which would
have taken a lot of space).
}
\details{
Wrapper around \code{plot_outcome_core}
}
\examples{

\dontrun{

root <- system.file("extdata", "sim-example", package = "speciomer")
plot_outcome_init(root)

}

}
\seealso{
\code{plot_outcome_core}
}
