% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outcome_simple.R
\name{plot_outcome_simple}
\alias{plot_outcome_simple}
\title{Plot a simplified simulation outcome}
\usage{
plot_outcome_simple(root, approx = NULL)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{approx}{Path to pass to \code{add_approx}. Leave \code{NULL} for no
overlay approximation plot.}
}
\value{
A patchwork
}
\description{
Equivalent of \code{plot_outcome_full} but without the plots of allele
frequencies, Fst, Qst and Cst through time (so only trait densities and
average mutational effects are shown).
}
\details{
Wrapper around \code{plot_outcome_core}
}
\note{
Locus-specific average mutational effects are subsampled every 1,000
generations to avoid crowding the plots.
}
\examples{

\dontrun{

root <- system.file("extdata", "sim-example", package = "speciomer")
plot_outcome_simple(root)

}

}
\seealso{
\code{plot_outcome_core}
}
