% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_locus.R
\name{plot_single_locus}
\alias{plot_single_locus}
\title{Plot a single locus against the rest of the genome}
\usage{
plot_single_locus(data, locus)
}
\arguments{
\item{data}{A data frame containing locus-specific data (see
\code{?read_loci})}

\item{locus}{Index of the locus to plot}
}
\value{
A ggplot
}
\description{
Plots the trajectory of one locus through time for some variables, on top
of a ribbon showing the 95 percent interval of trajectories across genes
coding for the same trait. Shows the allele frequency, Fst, Qst, Cst and
average mutational effect.
}
\details{
The color of the background ribbon corresponds to the trait encoded
by the locus.
}
\examples{

\dontrun{

root <- system.file("extdata", "sim-example", package = "speciomer")
data <- read_loci(root, c("freq", "Fst", "Qst", "Cst", "alpha"))
plot_single_locus(data, locus = 1)

}

}
\seealso{
\code{read_loci}
}
