% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_locus_extended.R
\name{plot_single_locus_extended}
\alias{plot_single_locus_extended}
\title{Plot a single locus in details}
\usage{
plot_single_locus_extended(genome_data, indiv_genomes, locus)
}
\arguments{
\item{genome_data}{Data frame containing locus-specific data (see
\code{?read_loci} and \code{?plot_single_locus})}

\item{indiv_genomes}{Data frame containing individual whole genome data
(see \code{?read_individual_genomes} and \code{plot_gene_regression})}

\item{locus}{Index of the locus to plot}
}
\value{
A patchwork
}
\description{
Plots several locus-specific variables through time for a specific locus,
together with information about the regression of genetic values against
genotypes across the population, at various time points.
}
\note{
The regression of genetic values against genotypes requires
whole-genome data at the individual-level, which may likely come from
a separate simulation. This is because it uses a lot of memory to record
individual whole genomes as often as one might record locus-specific
variables, which are summaries across the population.

See https://github.com/rscherrer/speciome for more details.
}
\examples{

root <- system.file("extdata", "sim-example", package = "speciomer")
root_indiv_genomes <- system.file(
  "extdata", "sim-indiv-genomes", package = "speciomer"
)
genome_data <- read_loci(root, c("freq", "Fst", "Qst", "Cst", "alpha"))
indiv_genomes <- read_individual_genomes(
  root_indiv_genomes, individual_variables = "ecotypes"
)

plot_single_locus_extended(genome_data, indiv_genomes, locus = 1)

}
\seealso{
\code{read_loci}, \code{read_individual_genomes},
\code{plot_single_locus}, \code{plot_gene_regression}
}
