% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_traits.R
\name{plot_traits}
\alias{plot_traits}
\title{Plot individual traits through time}
\usage{
plot_traits(data, burnin_bar = TRUE)
}
\arguments{
\item{data}{An individual-level dataset containing trait values for each
individual (see \code{?read_individuals}).}

\item{burnin_bar}{Whether to add a vertical bar at time point zero}
}
\value{
A ggplot
}
\description{
Plot the densities of each trait across the population and through time using
\code{geom_bin2d}.
}
\examples{

\dontrun{

root <- system.file("extdata", "sim-example", package = "speciomer")
data <- read_individuals(root, "individual_traits", ncol = 3)
plot_traits(data)

}

}
\seealso{
\code{read_individuals}
}
