% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_traits_from_root.R
\name{plot_traits_from_root}
\alias{plot_traits_from_root}
\title{Plot trait densities from root folder}
\usage{
plot_traits_from_root(root, burnin_bar = TRUE)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{burnin_bar}{Argument to be passed to \code{plot_traits}}
}
\value{
A ggplot
}
\description{
Wrapper around \code{plot_traits} that reads trait data and plots them from
the root folder of a simulation.
}
\examples{

\dontrun{

root <- system.file("extdata", "sim-example", package = "speciomer")
plot_traits_from_root(root)

}

}
\seealso{
\code{read_individuals}, \code{plot_traits}
}
