% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_architecture.R
\name{read_architecture}
\alias{read_architecture}
\title{Read genetic architecture}
\usage{
read_architecture(root)
}
\arguments{
\item{root}{Path to the simulation folder}
}
\value{
A named list of two tibbles.
}
\description{
Read the architecture file in a simulation folder and returns a pair of
tibbles, one for the loci and one for the edges.
}
\details{
The function looks for the file "architecture.txt" in the simulation folder
}
\examples{

root <- system.file("extdata", "sim-example", package = "speciomer")
read_architecture(root)

}
