% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bitset.R
\name{read_bitset}
\alias{read_bitset}
\title{Read a bitset}
\usage{
read_bitset(filename)
}
\arguments{
\item{filename}{Path to the file to read}
}
\value{
A vector of integers
}
\description{
Reads a binary file bit-wise and returns a vector of binary integers (0 or 1)
}
\examples{

root <- system.file("extdata", "sim-indiv-genomes", package = "speciomer")
read_bitset(paste0(root, "/individual_whole_genomes.dat"))

}
