% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_edges.R
\name{read_edges}
\alias{read_edges}
\title{Read edge-wise data}
\usage{
read_edges(root, variables, architecture = TRUE)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{variables}{Vector of names of variable to read (will be interpreted using \code{interpret_variable_names})}

\item{architecture}{Whether to append edge-wise genetic architecture parameters
(see \code{?read_architecture}).}
}
\value{
A tibble containing the simulation data
}
\description{
Wrapper around \code{read_speciome} for variables with one value per
edge per time point.
}
\details{
The function will read "paramlog.txt" in order to know
by how many loci to duplicate each time point.
}
\note{
The function will read "time.dat" as an extra variable, no need to
supply it.
}
\examples{

root <- system.file("extdata", "sim-example", package = "speciomer")
read_edges(root, "edge_corbreed")

}
\seealso{
\code{read_speciome}
}
