% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_indiv_loci.R
\name{read_indiv_loci}
\alias{read_indiv_loci}
\title{Read individual loci}
\usage{
read_indiv_loci(
  root,
  nloci = NULL,
  t = NULL,
  extra = NULL,
  freezerfile = "freezer.dat",
  locifile = "locivalues.dat",
  architecture = TRUE,
  archfile = "architecture.txt"
)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{nloci}{Number of loci. Guessed using `guess_nloci` if NULL.}

\item{t}{Optional time points to read. All if NULL.}

\item{extra}{Optional names of extra individual-wise variables to attach to
the output data frame (read using `read_pop`)}

\item{freezerfile}{File name of the freezer file}

\item{locifile}{File name of the loci values file}

\item{architecture}{Whether to add components of the genetic architecture}

\item{archfile}{Genetic architecture file name}
}
\value{
A data frame
}
\description{
Makes a data frame with alleles at each locus for each individual.
Can be very large
}
\details{
This function needs to read the file `population_size.dat`, so make
sure that population sizes are saved for the time points you want to load
whole genomes for. Otherwise, the function cannot assign time points to
each individual. This can be done when setting `tsave` and `tfreeze` in the
simulation (see the README for `speciome` on GitHub). If you do not have
these data and still want to read whole genomes (but not put them in a
tibble) you can use `read_bitset` instead.
}
\examples{

\dontrun{

root <- "data/example_1"

read_indiv_loci(root)

}

}
