% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_individual_genomes.R
\name{read_individual_genomes}
\alias{read_individual_genomes}
\title{Read individual whole genome data}
\usage{
read_individual_genomes(
  root,
  individual_variables = NULL,
  individual_ncols = NULL,
  locus_variables = NULL,
  locus_architecture = FALSE
)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{individual_variables, individual_ncols}{Variables to be passed to \code{read_individuals}}

\item{locus_variables, locus_architecture}{Variables to be passed to \code{read_loci}}
}
\value{
A tibble containing the simulation data
}
\description{
Reads individual genome data and assembles them into a tibble.
}
\details{
The files "individual_whole_genomes.dat" and "individual_locus_genvalues.dat"
are read and assembled into an individual-and-locus-wise tibble. The \code{individual_*}
and \code{locus_*} arguments allow to read extra individual-wise or locus-wise
variables, respectively, and append them to the tibble as metadata. Then, these
arguments are passed down to the functions \code{read_individuals} and
\code{read_loci}, respectively, as \code{variables}, \code{ncols} and \code{architecture}.
See \code{?read_individuals} and \code{?read_loci} for details.
}
\examples{

root <- system.file("extdata", "sim-indiv-genomes", package = "speciomer")
read_individual_genomes(root)
read_individual_genomes(root, "individual_ecotypes")
read_individual_genomes(root, locus_variables = "locus_Fst")

}
\seealso{
\code{read_individuals}, \code{read_loci}, \code{read_architecture}
}
