% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_individuals.R
\name{read_individuals}
\alias{read_individuals}
\title{Read individual-wise data}
\usage{
read_individuals(root, variables)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{variables}{Vector of names of variable to read}
}
\value{
A tibble containing the simulation data
}
\description{
Wrapper around \code{read_speciome} for variables with one value per
individual per time point.
}
\details{
The function will read "population_sizes.dat" in order to know
by how many individuals to duplicate each time point, as the number of
individuals may change from one generation to the next.
}
\note{
The function will read "time.dat" as an extra variable, no need to
supply it.
}
\examples{

}
\seealso{
`read_speciome`
}
