% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loci.R
\name{read_loci}
\alias{read_loci}
\title{Read locus-wise data}
\usage{
read_loci(root, variables, architecture = TRUE)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{variables}{Vector of names of variable to read (will be interpreted using \code{interpret_variable_names})}

\item{architecture}{Whether to append locus-wise genetic architecture parameters
(see \code{?read_architecture}).}
}
\value{
A tibble containing the simulation data
}
\description{
Wrapper around \code{read_speciome} for variables with one value per
locus per time point.
}
\details{
The function will read "paramlog.txt" in order to know
by how many loci to duplicate each time point.
}
\note{
The function will read "time.dat" as an extra variable, no need to
supply it.
}
\examples{

root <- system.file("extdata", "sim-example", package = "speciomer")
read_loci(root, "locus_Fst")

}
\seealso{
\code{read_speciome}
}
