% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_population.R
\name{read_population}
\alias{read_population}
\title{Read population-wide data}
\usage{
read_population(root, variables, ncols = NULL)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{variables}{Vector of names of variable to read}

\item{ncols}{Vector or list indicating how to split or duplicate each variable
when constructing the tibble. See \code{read_speciome}.}
}
\value{
A tibble containing the simulation data
}
\description{
Wrapper around \code{read_speciome} for variables with one value per time point.
}
\note{
The function will read "time.dat" as an extra variable, no need to
supply it.
}
\examples{

root <- system.file("extdata", "sim-example", package = "speciomer")
read_population(root, c("EI", "RI", "SI"))
read_population(root, c("EI", "trait_Fst"), ncols = c(1, 3))

}
\seealso{
\code{read_speciome}
}
