% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_traits.R
\name{read_traits}
\alias{read_traits}
\title{Read trait-wise data}
\usage{
read_traits(root, variables)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{variables}{Vector of names of variable to read (will be interpreted using \code{interpret_variable_names})}
}
\value{
A tibble containing the simulation data
}
\description{
Wrapper around \code{read_speciome} for variables with one value per
trait per time point.
}
\note{
The function will read "time.dat" as an extra variable, no need to
supply it.
}
\examples{

root <- system.file("extdata", "sim-example", package = "speciomer")
read_traits(root, "trait_Fst")

}
\seealso{
\code{read_speciome}
}
