% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_traits.R
\name{recode_traits}
\alias{recode_traits}
\title{Recode traits}
\usage{
recode_traits(x)
}
\arguments{
\item{x}{A vector of trait numbers}
}
\value{
A vector of factors
}
\description{
Change the names of the traits in a vector from 1, 2 and 3 to "Ecological",
"Mating" and "Neutral", respectively.
}
\examples{

recode_traits(1:3)

}
\seealso{
\code{forcats::fct_recode}
}
