% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_plural_colnames.R
\name{rm_plural_colnames}
\alias{rm_plural_colnames}
\title{Remove plurals}
\usage{
rm_plural_colnames(x)
}
\arguments{
\item{x}{A string vector}
}
\value{
A string vector
}
\description{
Removes "s" at the end of strings or just before digits at the end of strings
}
\examples{

rm_plural_colnames(c("traits1", "traits2", "traits3", "ecotypes"))

}
