% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_colors.R
\name{trait_colors}
\alias{trait_colors}
\title{Trait colors}
\usage{
trait_colors()
}
\value{
A vector of strings
}
\description{
Generates a vector of trait colors ("forestgreen", "goldenrod" and
"darkgray") named after their trait names ("Ecological", "Mating" and
"Neutral", respectively).
}
\examples{

trait_colors()

}
