% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_names.R
\name{trait_names}
\alias{trait_names}
\title{Trait names}
\usage{
trait_names()
}
\value{
A vector of strings
}
\description{
Generates a vector of trait names ("Ecological", "Mating" and "Neutral")
}
\examples{

trait_names()

}
