% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_numbers.R
\name{trait_numbers}
\alias{trait_numbers}
\title{Trait numbers}
\usage{
trait_numbers()
}
\value{
A vector of strings
}
\description{
Generates a vector of trait numbers (1, 2 and 3) named after their trait
names ("Ecological", "Mating" and "Neutral", respectively).
}
\examples{

trait_numbers()

}
