% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_individuals.R
\name{read_individuals}
\alias{read_individuals}
\title{Read individual-wise data}
\usage{
read_individuals(root, variables, ncols = NULL)
}
\arguments{
\item{root}{Path to the simulation folder}

\item{variables}{Vector of names of variable to read}

\item{ncols}{Vector or list indicating how to split or duplicate each variable
when constructing the tibble. See \code{read_speciome}.}
}
\value{
A tibble containing the simulation data
}
\description{
Wrapper around \code{read_speciome} for variables with one value per
individual per time point.
}
\details{
The function will read "population_sizes.dat" in order to know
by how many individuals to duplicate each time point, as the number of
individuals may change from one generation to the next.
}
\note{
The function will read "time.dat" as an extra variable, no need to
supply it.
}
\examples{

root <- system.file("extdata", "sim-example", package = "speciomer")
read_individuals(root, "individual_ecotypes")
read_individuals(root, c("individual_traits", "individual_ecotypes"), ncols = c(3, 1))

}
\seealso{
\code{read_speciome}
}
