% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_speciome.R
\name{read_speciome}
\alias{read_speciome}
\title{Read speciome simulation data}
\usage{
read_speciome(root, variables, ncols = rep(1, length(variables)))
}
\arguments{
\item{root}{Path to the simulation folder}

\item{variables}{Vector of names of variable to read (e.g. \code{c("time", "trait_Fst")})}

\item{ncols}{Vector or list indicating how to split or duplicate each variable
when constructing the tibble. There must be one element per variable.
See \code{?as_tibble_speciome} for details.}
}
\value{
A tibble containing the simulation data
}
\description{
Combine data from one simulation, saved as binary files, into a single
tibble.
}
\details{
Each variable is read as a vector and reshaped into a tibble by being
passed, with its corresponding \code{ncols}, to \code{as_tibble_speciome}.
The resulting tibbles for all variables are then bound together by column.
}
\note{
Do not provide the extension of the data files in \code{variables} (".dat").
}
\examples{

root <- system.file("extdata", "sim-example", package = "speciomer")
read_speciome(root, "time")
read_speciome(root, c("time", "EI"))
read_speciome(root, c("time", "trait_Fst"), ncol = c(1, 3))
read_speciome(root, c("time", "trait_Fst"), ncol = c(-3, 1))

}
\seealso{
\code{as_tibble_speciome}
}
